Create Analysis of Weather Data in Australia README.md
The Rain in Australia dataset is a dataset from Kaggle. This dataset contains about 10 years of daily weather observations from 49 locations across Australia. The weatherAUS.csv file contains 145460 rows and 23 columns. The variables include both quantitative variables and categorical variables. In order to understand the Rain pattern in Australia and predict if it is going to rain tomorrow in Australia, we focused on three geographically separated locations, Perth, Melbourne Airport and Sydney Airport.